pyproject.toml
[project]
name = "mcp-ai-webscrapping"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=3.1.0",
    "groq>=1.0.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
    "trafilatura>=2.0.0",
]
